Convert the Billing project to ASP.NET
